requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-dotenv
requests
langchain-ollama
langchain-openai
